Moved CurrencyQuery class
