+++ app/src/main/java/com/example/trippe/ui/currency/CurrencyFragment.java
@@ -3,7 +3,6 @@ package com.example.trippe.ui.currency;
 import android.content.Context;
 import android.graphics.Color;
 import android.os.Bundle;
-import android.text.Layout;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -21,7 +20,6 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
 
-import com.example.trippe.CurrencyQuery;
 import com.example.trippe.R;
 import com.jjoe64.graphview.GraphView;
 import com.jjoe64.graphview.series.DataPoint;
